+++ app/src/main/AndroidManifest.xml
@@ -15,10 +15,10 @@
         android:theme="@style/Theme.Yumyum"
         android:usesCleartextTraffic="true">
         <activity
-            android:name=".presentation.ui.withdraw.WithdrawActivity"
+            android:name="com.doctor.yumyum.presentation.ui.mypage.withdraw.WithdrawActivity"
             android:exported="true" />
         <activity
-            android:name=".presentation.ui.myinfo.MyInfoActivity"
+            android:name="com.doctor.yumyum.presentation.ui.mypage.myinfo.MyInfoActivity"
             android:exported="true" />
         <activity
             android:name=".presentation.ui.search.hashtag.SearchHashtagActivity"
