+++ app/src/main/java/com/doctor/yumyum/presentation/ui/mypage/myinfo/MyInfoActivity.kt
@@ -1,4 +1,4 @@
-package com.doctor.yumyum.presentation.ui.myinfo
+package com.doctor.yumyum.presentation.ui.mypage.myinfo
 
 import android.content.Intent
 import android.os.Bundle
@@ -6,7 +6,7 @@ import com.doctor.yumyum.R
 import com.doctor.yumyum.common.base.BaseActivity
 import com.doctor.yumyum.databinding.ActivityMyInfoBinding
 import com.doctor.yumyum.presentation.ui.nickname.NicknameActivity
-import com.doctor.yumyum.presentation.ui.withdraw.WithdrawActivity
+import com.doctor.yumyum.presentation.ui.mypage.withdraw.WithdrawActivity
 
 class MyInfoActivity : BaseActivity<ActivityMyInfoBinding>(R.layout.activity_my_info) {
     override fun onCreate(savedInstanceState: Bundle?) {
