+++ app/src/main/res/layout/activity_my_info.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.ui.myinfo.MyInfoActivity">
+        tools:context=".presentation.ui.mypage.myinfo.MyInfoActivity">
 
         <include
             android:id="@+id/my_info_appbar"
