+++ app/src/main/res/layout/activity_withdraw.xml
@@ -14,7 +14,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.ui.withdraw.WithdrawActivity">
+        tools:context=".presentation.ui.mypage.withdraw.WithdrawActivity">
 
         <include
             android:id="@+id/withdraw_appbar"
